3_Neural Network from scratch/deep-learning-pipeline-lecture3/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-learning-pipeline"
version = "1.0.0"
description = "Deep Learning Pipeline for Lecture 3"
readme = "README.md"
authors = [
    {name = "Deep Learning Course"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "scikit-learn>=1.0.0",
    "tqdm>=4.62.0",
    "PyYAML>=6.0",
    "gradio>=3.0.0",
    "pandas>=1.3.0",
    "networkx>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "jupyter>=1.0.0",
]

[project.scripts]
dl-pipeline = "scripts.run_pipeline:main"
dl-train = "scripts.train_model:main"
dl-download = "scripts.download_data:main"
dl-ui = "scripts.launch_ui:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']